lang

+++ app/Http/Controllers/Api/ProductController.php
@@ -2,21 +2,41 @@
 
 namespace App\Http\Controllers\Api;
 
-use App\Http\Controllers\Controller;
 use App\Models\Product;
 use Illuminate\Http\Request;
+use App\Http\Controllers\Controller;
+use App\Http\Resources\ProductResource;
 use Illuminate\Support\Facades\Response;
+use Illuminate\Routing\Controllers\Middleware;
+use Illuminate\Routing\Controllers\HasMiddleware;
+use Illuminate\Support\Facades\Auth;
 
-class ProductController extends Controller
+class ProductController extends Controller implements HasMiddleware
 {
-    /**
+
+    // public function __construct()
+    // {
+    //     $this->middleware('auth:sanctum')->except(['index','show']);
+    // }
+
+   /**
+     * Get the middleware that should be assigned to the controller.
+     */
+    public static function middleware(): array
+    {
+        return [
+            new Middleware('auth:sanctum', except:['index','show']),
+        ];
+    }
+        /**
      * Display a listing of the resource.
      */
-    public function index(Request $request)      //laravel outomatec convert model collection to json
+    public function index(Request $request)      //laravel outmatchen convert model collection to json
     {
-        return Product::active()->Filter($request->query())
-        ->with('category:id,name','store:id,name','tags:id,name') //respons id and name onle id use not matcheng laravel beetwean product name and category name
+         $products= Product::active()->Filter($request->query())
+        //->with('category:id,name','store:id,name','tags:id,name') //respons id and name onle id use not matcheng laravel beetwean product name and category name
         ->paginate();
+        return ProductResource::collection($products);
     }
 
     /**
@@ -27,6 +47,12 @@ class ProductController extends Controller
         $validated = $request->validate(Product::rules());
 
         $product = Product::create($validated);
+
+        $user=$request->user();
+        if(!$user->tokenCan('products.create'))
+        {
+            abort(403,'Not allowed');
+        }
 
         return  response()->json($product, 201, [
             'Loction'=> route('products.show',$product->id), // Response facade becose header 
@@ -39,7 +65,8 @@ class ProductController extends Controller
      */
     public function show(Product $product)
     {
-        return $product->load('category:id,name','store:id,name','tags:id,name');  //return object  //with return bulider
+        return new ProductResource($product);
+        //->load('category:id,name','store:id,name','tags:id,name');  //return object  //with return bulider
         
     }
 
@@ -49,6 +76,12 @@ class ProductController extends Controller
     public function update(Request $request,Product $product)
     {
         $validated = $request->validate(Product::rules());
+
+        $user=$request->user();
+        if(!$user->tokenCan('products.update'))
+        {
+            abort(403,'Not allowed');
+        }
 
         $product->update($validated);
         return  response()->json($product, 201, [
@@ -62,6 +95,15 @@ class ProductController extends Controller
      */
     public function destroy(string $id)
     {
+
+        $user=Auth::user();
+        if(!$user->tokenCan('products.delete'))
+        {
+            return response([
+                'massege'=> 'Not allowed',
+            ],403);
+            // abort(403,'Not allowed'); return massege and Exception
+        }
        Product::destroy($id);
        return [                                         //laravel automatec array to json
         'massege'=>'Product Deleted Succsesfully'
